Feature/modify question

+++ src/main/java/com/example/initbackend/question/dto/IssueQuestionIdRequestDto.java
@@ -9,14 +9,14 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class IssueQuestionIdRequestDto {
 
-    private Integer userId;
+    private Long userId;
 
     @Builder
-    public IssueQuestionIdRequestDto(Integer userId) {
+    public IssueQuestionIdRequestDto(Long userId) {
         this.userId = userId;
     }
 
-    public static Question toEntity(Integer userId){
+    public static Question toEntity(Long userId){
         return Question.builder()
                 .user_id(userId)
                 .build();
